pyproject.toml
[project]
name = "directdnsonly"
version = "1.0.9"
description = "DNS Management System - DirectAdmin to multiple backends"
authors = [
    {name = "Aaron Guise",email = "aaron@guise.net.nz"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "vyper-config (>=1.2.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "persist-queue (>=1.0.0,<2.0.0)",
    "cherrypy (>=18.10.0,<19.0.0)",
    "sqlalchemy (<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]

[tool.poetry]
package-mode = true

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pyinstaller = "^6.13.0"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
